scrapers/utils/requirements.txt
psycopg2
pytesseract
Pillow